requirements.txt
streamlit==1.38.0
streamlit-folium==0.23.2
folium==0.17.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
bcrypt==4.2.0
opencv-python==4.10.0.84
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.2
protobuf==4.25.3
